Edit workshop801 success !!

+++ WithdrawalDemoSystem/WithdrawalDemoSystem/Withdrawal.vb
@@ -1,18 +1,18 @@
 ﻿Public Class Withdrawal
-    Private _Withdraw, _Balance As Integer
+    Private _Withdraw, _Balance As Double
     Public Sub New(withdraw)
         _Withdraw = withdraw
     End Sub
 
-    Public Function Foundcash1000() As Integer
+    Public Function Foundcash1000() As Double
         Return _Withdraw \ 1000
     End Function
 
-    Public Function Foundcash500() As Integer
+    Public Function Foundcash500() As Double
         Return (_Withdraw Mod 1000) \ 500
     End Function
 
-    Public Function Foundcash100() As Integer
+    Public Function Foundcash100() As Double
         Return ((_Withdraw Mod 1000) Mod 500) \ 100
     End Function
